+++ src/background_scripts/extractors/mangahere.js
@@ -13,7 +13,7 @@ export const urlRegex = /(?:https*:\/\/[\w.]*)*(mangahere)\.\w{2,3}\/manga\/([\w
  * @returns A string representing the chapter reference extracted from the URL or the defaultValue.
  */
 export function getChapterReference(url, defaultValue) {
-  const m = /[\w:/.]+\/manga\/\w+\/([\dc.]+)\/(?:\d+\.html)*/.exec(url);
+  const m = /[\w:/.]+\/manga\/\w+\/([\dvc/.]+)\/(?:\d+\.html)*/.exec(url);
   return (m) ? { id: m[1] } : defaultValue;
 }
 
@@ -24,15 +24,27 @@ export function getChapterReference(url, defaultValue) {
  * @returns String representing the manga cover URL.
  */
 export function getMangaCover(response, url) {
-  if (!response || typeof response !== 'object') {
-    throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, url);
-  }
-
   // Get manga image URL from response
   const imageDom = response.evaluate('//div[contains(@class,"manga_detail_top clearfix")]/img', response, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
   if (!imageDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_COVER, url);
 
   return imageDom.singleNodeValue.getAttribute('src');
+}
+
+/**
+ * Returns the Manga status (Ongoing|Completed) from the manga main page.
+ * @param {object} response The Fetch API response object.
+ * @param {string} url Optional. The URL to print with errors.
+ * @returns {boolean} True if the series is stated as 'Completed' and False otherwise.
+ */
+export function getMangaStatus(response, url) {
+  const imageDom = response.evaluate('//ul[@class="detail_topText"]/li[7]', response, null, XPathResult.FIRST_ORDERED_NODE_TYPE, null);
+  if (!imageDom.singleNodeValue) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  const statusStr = /Status:(\w+)/.exec(imageDom.singleNodeValue.textContent);
+  if (!statusStr) throw FoxyError(ErrorCode.NO_MANGA_STATUS, url);
+
+  return (statusStr[1] === 'Completed');
 }
 
 /**
@@ -45,7 +57,7 @@ function getChapterList(mangaSid, mangaUrl, chapterList = []) {
   const url = `http://www.mangahere.cc/get_chapters${mangaSid}.js`;
 
   return HttpFetch(url, (response) => {
-    const regex = /\["(.+)","[\w./"+]+\/([c\d.]+)\/"\]/g;
+    const regex = /\["(.+)","[\w./"+]+"\/([vc/\d.]+)\/"\]/g;
 
     const body = response.body.innerHTML;
     if (!body) return chapterList;
@@ -95,8 +107,9 @@ export function getMangaInfo(url) {
 
     const name = titleDom.singleNodeValue.getAttribute('content');
 
-    // Get manga SID and image URL from response
+    // Get image URL and status from response
     const imageUrl = getMangaCover(response, mangaUrl);
+    const status = getMangaStatus(response, mangaUrl);
 
     m = /\S+\/manga\/(\d+)\S+/.exec(imageUrl);
     if (!m) throw FoxyError(ErrorCode.NO_MANGA_SID, mangaUrl);
@@ -115,6 +128,7 @@ export function getMangaInfo(url) {
         reference: mangaReference,
         url: mangaUrl,
         cover: imageUrl,
+        status,
         last_update: new Date(),
         chapter_list: chapterList,
       };
@@ -148,4 +162,28 @@ export async function updateChapters(manga) {
   } catch (err) {
     throw err;
   }
+}
+
+/**
+ * Returns the manga information
+ * @param {object} manga The manga to update.
+ * @returns {Promise} A promise which resolves to the manga object retrieved.
+ */
+export async function updateMetadata(manga) {
+  if (!manga || !manga.url) {
+    throw new TypeError(`updateMetadata() argument is invalid: ${JSON.stringify(manga)}`);
+  }
+
+  return HttpFetch(manga.url, async (response) => {
+    if (!response || typeof response !== 'object') {
+      throw FoxyError(ErrorCode.RESPONSE_NOT_HTML, manga.url);
+    }
+
+    const update = {
+      cover: getMangaCover(response, manga.url),
+      status: getMangaStatus(response, manga.url),
+    };
+
+    return update;
+  });
 }