+++ test/kissmanga.js
@@ -5,7 +5,7 @@ import * as KissManga from '../src/background_scripts/extractors/kissmanga';
 
 const should = chai.should();
 
-describe.only('KissManga', () => {
+describe('KissManga', () => {
   before(() => {
     // Setting DOMParser global to JSDOM
     global.DOMParser = window.DOMParser;
@@ -45,10 +45,6 @@ describe.only('KissManga', () => {
       parser = new DOMParser();
     });
 
-    it('should throw error if response is not a DOM object', () => {
-      (KissManga.getMangaCover).should.throw(Error, /Foxy Error #300/);
-    });
-
     it('should throw error if no <link rel="image_src" /> could be retrieved from response body', () => {
       const response = parser.parseFromString(`
         <!DOCTYPE html>
@@ -57,7 +53,7 @@ describe.only('KissManga', () => {
           </head>
           <body>
           </body>
-        <html>`, 'text/html');
+        </html>`, 'text/html');
 
       const fn = () => { KissManga.getMangaCover(response); };
 
@@ -74,12 +70,83 @@ describe.only('KissManga', () => {
           <body>
             <div><img src="http://kissmanga.com/Uploads/Etc/10-11-2011/1.jpg"/></div>
           </body>
-        <html>`, 'text/html');
+        </html>`, 'text/html');
 
       const cover = KissManga.getMangaCover(response);
 
       should.exist(cover);
       cover.should.be.equal('http://kissmanga.com/Uploads/Etc/10-11-2011/6554082i76520.jpg');
+    });
+  });
+
+  // Test for #getMangaStatus()
+  describe('#getMangaStatus()', () => {
+    let parser;
+    before(() => {
+      parser = new DOMParser();
+    });
+
+    it('should throw error if no "Status: X" could be retrieved from response body', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+          </body>
+        </html>`, 'text/html');
+
+      const fn = () => { KissManga.getMangaStatus(response); };
+
+      (fn).should.throw(Error, /Foxy Error #113/);
+    });
+
+    it('should return true if status is Completed', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+            <div id="leftside">
+              <div>
+                <div class="barContent">
+                  <div></div>
+                  <div>Status: Completed</div>
+                </div>
+              </div>
+            </div>
+          </body>
+        </html>`, 'text/html');
+
+      const status = KissManga.getMangaStatus(response);
+
+      should.exist(status);
+      status.should.be.equal(true);
+    });
+
+    it('should return false if status is Ongoing', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+            <div id="leftside">
+              <div>
+                <div class="barContent">
+                  <div></div>
+                  <div>Status: Ongoing</div>
+                </div>
+              </div>
+            </div>
+          </body>
+        </html>`, 'text/html');
+
+      const status = KissManga.getMangaStatus(response);
+
+      should.exist(status);
+      status.should.be.equal(false);
     });
   });
 
@@ -185,6 +252,36 @@ describe.only('KissManga', () => {
         });
     });
 
+    it('should reject promise on cover error', () => {
+      window.fetch.callsFake(MockFetch.ok(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+            <meta name="description" content="Read I Can&#39;t Sleep Alone manga online free and high quality." />
+          </head>
+          <body>
+          <div id="leftside">
+          </div>
+          </body>
+        </html>`));
+
+      const promise = KissManga.updateMetadata({
+        url: 'http://kissmanga.com/Manga/I-Can-t-Sleep-Alone',
+      });
+
+      return promise
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('Foxy Error #104');
+          err.params.should.have.string('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone');
+        });
+    });
+
     it('should return the correct manga object structure for chapter page', () => {
       const res = `
         <!DOCTYPE html>
@@ -195,6 +292,10 @@ describe.only('KissManga', () => {
           </head>
           <body>
           <div id="leftside">
+            <div class="barContent">
+              <div></div>
+              <div>Status: Ongoing</div>
+            </div>
             <table>
             <tr>
               <td><a href="/Manga/One-Piece/One-Piece---410?id=319813">One Piece 410</a></td>
@@ -205,7 +306,7 @@ describe.only('KissManga', () => {
             </table>
           </div>
           </body>
-        <html>`;
+        </html>`;
 
       window.fetch
         .withArgs('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone').callsFake(MockFetch.ok(res));
@@ -222,6 +323,7 @@ describe.only('KissManga', () => {
           manga.should.have.property('reference').equal('I-Can-t-Sleep-Alone');
           manga.should.have.property('url').equal('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone');
           manga.should.have.property('cover').equal('http://kissmanga.com/Uploads/Etc/10-11-2011/6554082i76520.jpg');
+          manga.should.have.property('status').equal(false);
           manga.should.have.property('last_update');
           manga.should.have.property('chapter_list').with.lengthOf(2);
         }).catch((err) => {
@@ -239,6 +341,10 @@ describe.only('KissManga', () => {
           </head>
           <body>
           <div id="leftside">
+            <div class="barContent">
+              <div></div>
+              <div>Status: Completed</div>
+            </div>
             <table>
             <tr>
               <td><a href="/Manga/One-Piece/One-Piece---410?id=319813">One Piece 410</a></td>
@@ -249,7 +355,7 @@ describe.only('KissManga', () => {
             </table>
           </div>
           </body>
-        <html>`;
+        </html>`;
 
       window.fetch
         .withArgs('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone').callsFake(MockFetch.ok(res));
@@ -266,6 +372,7 @@ describe.only('KissManga', () => {
           manga.should.have.property('reference').equal('I-Can-t-Sleep-Alone');
           manga.should.have.property('url').equal('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone');
           manga.should.have.property('cover').equal('http://kissmanga.com/Uploads/Etc/10-11-2011/6554082i76520.jpg');
+          manga.should.have.property('status').equal(true);
           manga.should.have.property('last_update');
           manga.should.have.property('chapter_list').with.lengthOf(2);
         }).catch((err) => {
@@ -386,6 +493,109 @@ describe.only('KissManga', () => {
           });
         })
         .catch((err) => {
+          should.not.exist(err);
+        });
+    });
+  });
+
+  // Test for #updateMetadata()
+  describe('#updateMetadata()', () => {
+    // Sinon sandbox for Fetch API
+    beforeEach(() => {
+      MockFetch.init();
+    });
+    afterEach(() => {
+      MockFetch.restore();
+    });
+
+    it('should reject promise if manga object is invalid', () => {
+      return KissManga.updateMetadata({ })
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(0);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('updateMetadata() argument is invalid');
+        });
+    });
+
+    it('should reject promise on Error 404', () => {
+      const promise = KissManga.updateMetadata({
+        url: 'http://kissmanga.com/Manga/I-Can-t-Sleep-Alone',
+      });
+
+      return promise
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('Foxy Error #302');
+          err.params.should.have.string('Not Found');
+        });
+    });
+
+    it('should reject promise on cover error', () => {
+      window.fetch.callsFake(MockFetch.ok(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+          </body>
+        </html>`));
+
+      const promise = KissManga.updateMetadata({
+        url: 'http://kissmanga.com/Manga/I-Can-t-Sleep-Alone',
+      });
+
+      return promise
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('Foxy Error #104');
+          err.params.should.have.string('http://kissmanga.com/Manga/I-Can-t-Sleep-Alone');
+        });
+    });
+
+    it('should resolve to the correct update object', () => {
+      window.fetch.callsFake(MockFetch.ok(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+            <link rel="image_src" href="http://kissmanga.com/Uploads/Etc/10-11-2011/6554082i76520.jpg" />
+          </head>
+          <body>
+            <div id="leftside">
+              <div class="barContent">
+                <div></div>
+                <div>Status: Completed</div>
+              </div>
+            </div>
+          </body>
+        </html>`));
+
+      const promise = KissManga.updateMetadata({
+        url: 'http://kissmanga.com/Manga/I-Can-t-Sleep-Alone',
+      });
+
+      return promise
+        .then((data) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(data);
+          data.should.be.an('object');
+          data.should.have.property('cover').equal('http://kissmanga.com/Uploads/Etc/10-11-2011/6554082i76520.jpg');
+          data.should.have.property('status').equal(true);
+        }).catch((err) => {
           should.not.exist(err);
         });
     });