+++ test/mangafox.js
@@ -46,10 +46,6 @@ describe('Mangafox', () => {
       parser = new DOMParser();
     });
 
-    it('should throw error if response is not a DOM object', () => {
-      (Mangafox.getMangaCover).should.throw(Error, /Foxy Error #300/);
-    });
-
     it('should throw error if no class="cover" could be retrieved from response body', () => {
       const response = parser.parseFromString(`
         <!DOCTYPE html>
@@ -58,7 +54,7 @@ describe('Mangafox', () => {
           </head>
           <body>
           </body>
-        <html>`, 'text/html');
+        </html>`, 'text/html');
 
       const fn = () => { Mangafox.getMangaCover(response); };
 
@@ -76,12 +72,79 @@ describe('Mangafox', () => {
           <body>
             <div class="cover"><img src="https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg"/></div>
           </body>
-        <html>`, 'text/html');
+        </html>`, 'text/html');
 
       const cover = Mangafox.getMangaCover(response);
 
       should.exist(cover);
       cover.should.be.equal('https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg');
+    });
+  });
+
+  // Test for #getMangaStatus()
+  describe('#getMangaStatus()', () => {
+    let parser;
+    before(() => {
+      parser = new DOMParser();
+    });
+
+    it('should throw error if no <div class="data"></div> could be retrieved from response body', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+          </body>
+        </html>`, 'text/html');
+
+      const fn = () => { Mangafox.getMangaStatus(response); };
+
+      (fn).should.throw(Error, /Foxy Error #113/);
+    });
+
+    it('should return true if status is Completed', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+            <div id="series_info">
+              <div class="data">
+                <h5>Status:</h5>
+                <span> Completed</span>
+              </div>
+            </div>
+          </body>
+        </html>`, 'text/html');
+
+      const status = Mangafox.getMangaStatus(response);
+
+      should.exist(status);
+      status.should.be.equal(true);
+    });
+
+    it('should return false if status is Ongoing', () => {
+      const response = parser.parseFromString(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+            <div id="series_info">
+              <div class="data">
+                <h5>Status:</h5>
+                <span>Ongoing, Next</span>
+              </div>
+            </div>
+          </body>
+        </html>`, 'text/html');
+
+      const status = Mangafox.getMangaStatus(response);
+
+      should.exist(status);
+      status.should.be.equal(false);
     });
   });
 
@@ -241,8 +304,14 @@ describe('Mangafox', () => {
           </head>
           <body>
             <div class="cover"><img src="https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg"/></div>
+            <div id="series_info">
+              <div class="data">
+                <h5>Status:</h5>
+                <span> Completed</span>
+              </div>
+            </div>
           </body>
-        <html>`;
+        </html>`;
       const res2 = `
         ["Title 1","v01/c001"],
         ["Title 2","v01/c002.5"]`;
@@ -263,6 +332,7 @@ describe('Mangafox', () => {
           manga.should.have.property('reference').equal('12_manga');
           manga.should.have.property('url').equal('http://fanfox.net/manga/12_manga/');
           manga.should.have.property('cover').equal('https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg');
+          manga.should.have.property('status').equal(true);
           manga.should.have.property('last_update');
           manga.should.have.property('chapter_list').with.lengthOf(2);
         }).catch((err) => {
@@ -280,8 +350,14 @@ describe('Mangafox', () => {
           </head>
           <body>
             <div class="cover"><img src="https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg"/></div>
+            <div id="series_info">
+              <div class="data">
+                <h5>Status:</h5>
+                <span>Ongoing,</span>
+              </div>
+            </div>
           </body>
-        <html>`;
+        </html>`;
       const res2 = `
         ["Title 1","v01/c001"],
         ["Title 2","v01/c002.5"]`;
@@ -302,6 +378,7 @@ describe('Mangafox', () => {
           manga.should.have.property('reference').equal('12_manga');
           manga.should.have.property('url').equal('http://fanfox.net/manga/12_manga/');
           manga.should.have.property('cover').equal('https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg');
+          manga.should.have.property('status').equal(false);
           manga.should.have.property('last_update');
           manga.should.have.property('chapter_list').with.lengthOf(2);
         }).catch((err) => {
@@ -407,6 +484,109 @@ describe('Mangafox', () => {
           });
         })
         .catch((err) => {
+          should.not.exist(err);
+        });
+    });
+  });
+
+  // Test for #updateMetadata()
+  describe('#updateMetadata()', () => {
+    // Sinon sandbox for Fetch API
+    beforeEach(() => {
+      MockFetch.init();
+    });
+    afterEach(() => {
+      MockFetch.restore();
+    });
+
+    it('should reject promise if manga object is invalid', () => {
+      return Mangafox.updateMetadata({ })
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(0);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('updateMetadata() argument is invalid');
+        });
+    });
+
+    it('should reject promise on Error 404', () => {
+      const promise = Mangafox.updateMetadata({
+        url: 'http://fanfox.net/manga/12_manga/',
+      });
+
+      return promise
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('Foxy Error #302');
+          err.params.should.have.string('Not Found');
+        });
+    });
+
+    it('should reject promise on cover error', () => {
+      window.fetch.callsFake(MockFetch.ok(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+          </body>
+        </html>`));
+
+      const promise = Mangafox.updateMetadata({
+        url: 'http://fanfox.net/manga/12_manga/',
+      });
+
+      return promise
+        .then((data) => {
+          should.not.exist(data);
+        }).catch((err) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(err);
+          err.should.be.an('error');
+          err.message.should.have.string('Foxy Error #104');
+          err.params.should.have.string('http://fanfox.net/manga/12_manga/');
+        });
+    });
+
+    it('should resolve to the correct update object', () => {
+      window.fetch.callsFake(MockFetch.ok(`
+        <!DOCTYPE html>
+        <html>
+          <head>
+          </head>
+          <body>
+            <div class="cover"><img src="https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg"/></div>
+            <div id="series_info">
+              <div class="data">
+                <h5>Status:</h5>
+                <span>Completed</span>
+              </div>
+            </div>
+          </body>
+        </html>`));
+
+      const promise = Mangafox.updateMetadata({
+        url: 'http://fanfox.net/manga/12_manga/',
+      });
+
+      return promise
+        .then((data) => {
+          MockFetch.callCount().should.be.equal(1);
+
+          should.exist(data);
+          data.should.be.an('object');
+          data.should.have.property('cover').equal('https://lmfcdn.secure.footprint.net/store/manga/5035/cover2.jpg');
+          data.should.have.property('status').equal(true);
+        }).catch((err) => {
           should.not.exist(err);
         });
     });